English revision
Revision of most of the text (code, some parts of tables and SDK docs were not revised). Most corrections were on orthography or gramatics, but there were some (less than 10) changes on full sentences. Also, there was a major exclusion of a duplicated and incomplete block in the Tunnel section.